添加transactioncontroller测试类
