+++ Software_Group76-Remaining part/FinanceApp/data/transactions.csv
@@ -1,3 +1,3 @@
 日期,分类,金额,备注
-2025-05-17,Food,10.00,Lunch
-2025-05-17,Transportation,5.00,Bus
+2025-05-17,Food,10.00,Breakfast
+2025-05-17,Shopping,50.00,Clothes
